requirements.txt
# Python Dependencies
fastapi>=0.104.1
uvicorn>=0.24.0
langchain>=0.1.0
langchain-openai>=0.0.5
openai>=1.3.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
httpx>=0.25.2
aiofiles>=23.2.0
pillow>=10.1.0
requests>=2.31.0
psutil>=5.9.6
pytest>=7.4.3
pytest-asyncio>=0.21.1
